feat(Front): Improve time filtering performance + Clean old code
