+++ front/src/components/Colors/ColorsAlphas.vue
@@ -3,7 +3,7 @@ import type {InputNumberProps} from 'naive-ui';
 import {NInputNumber, NTooltip} from 'naive-ui';
 
 import {alphaHighRef, alphaLowRef} from '.././Scatter/useScatterColorScale';
-import {scatterReadyRef} from '../Scatter/useScatterStatus';
+import {scatterLoadingRef} from '../Scatter/useScatterLoading';
 
 type InputNumberThemeOverrides = NonNullable<
   InputNumberProps['themeOverrides']
@@ -26,7 +26,7 @@ const inputNumberThemeOverrides: InputNumberThemeOverrides = {
     <template #trigger>
       <n-input-number
         v-model:value="alphaLowRef.value"
-        :disabled="!scatterReadyRef.value"
+        :disabled="scatterLoadingRef.value"
         :theme-overrides="inputNumberThemeOverrides"
         class="input"
         max="1"
@@ -45,7 +45,7 @@ const inputNumberThemeOverrides: InputNumberThemeOverrides = {
     <template #trigger>
       <n-input-number
         v-model:value="alphaHighRef.value"
-        :disabled="!scatterReadyRef.value"
+        :disabled="scatterLoadingRef.value"
         :theme-overrides="inputNumberThemeOverrides"
         class="input"
         max="1"
