+++ front/src/components/Help/Help.vue
@@ -42,14 +42,10 @@ const shortcuts: Shortcut[] = [
   {key: KeyboardShortcut.selection, name: 'Selection'},
   {key: KeyboardShortcut.trajectories, name: 'Trajectories'},
   {key: KeyboardShortcut.colors, name: 'Colors'},
-  {key: KeyboardShortcut.queries, name: 'Query'},
   {key: KeyboardShortcut.time, name: 'Time'},
   {key: KeyboardShortcut.meta, name: 'Meta'},
   {key: KeyboardShortcut.audio, name: 'Audio'},
   {key: KeyboardShortcut.details, name: 'Details'},
-  {key: KeyboardShortcut.volumes, name: 'Volumes'},
-  {key: KeyboardShortcut.matrices, name: 'Matrices'},
-  {key: KeyboardShortcut.pairings, name: 'Pairings'},
   {key: KeyboardShortcut.timePlayPause, name: 'Time: Play / Pause'},
   {key: KeyboardShortcut.timeBackward, name: 'Time: Next'},
   {key: KeyboardShortcut.timeForward, name: 'Time: Previous'},
