--- front/src/components/Queries/queryStore.ts
@@ -1,11 +1,0 @@
-import {reactive} from 'vue';
-
-export interface QueryStore {
-  matches: string[];
-  query: string;
-}
-
-export const queryStore = reactive<QueryStore>({
-  matches: [],
-  query: '',
-});
