+++ front/src/components/Scatter/useScatterFeatures.ts
@@ -9,6 +9,7 @@ import {pointsFilteredByTimeRef} from './useScatterFilterTime';
 
 export function useScatterFeatures() {
   const traceFeatures = (): Data[] => {
+    console.log('traceFeatures');
     if (
       labelsPropertiesRef.value === null ||
       reducedFeaturesRef.value === null ||
