+++ front/src/components/Scatter/useScatterLoading.ts
@@ -1,0 +1,9 @@
+import {reactive} from 'vue';
+
+interface ScatterLoadingRef {
+  value: boolean;
+}
+
+export const scatterLoadingRef = reactive<ScatterLoadingRef>({
+  value: false,
+});
